+++ install.ps1
@@ -6,8 +6,10 @@
 # If unable to execute due to policy rules, run
 # Set-ExecutionPolicy RemoteSigned -Scope CurrentUser.
 
-# Exit immediately if a PowerShell Cmdlet encounters an error.
+# Exit immediately if a PowerShell cmdlet encounters an error.
 $ErrorActionPreference = 'Stop'
+# Disable progress bar for cmdlets.
+$ProgressPreference = 'SilentlyContinue'
 # Exit immediately when an native executable encounters an error.
 $PSNativeCommandUseErrorActionPreference = $True
 
@@ -27,18 +29,10 @@ Options:
 '@
 }
 
-# Downloads file to destination efficiently.
-Function DownloadFile($SrcURL, $DstFile) {
-    # The progress bar updates every byte, which makes downloads slow. See
-    # https://stackoverflow.com/a/43477248 for an explanation.
-    $ProgressPreference = 'SilentlyContinue'
-    Invoke-WebRequest -UseBasicParsing -OutFile "$DstFile" -Uri "$SrcURL"
-}
-
 # Print error message and exit script with usage error code.
 Function ErrorUsage($Message) {
-    Write-Error "Error: $Message"
-    Write-Error "Run 'install --help' for usage"
+    Write-Output "error: $Message"
+    Write-Output "Run 'install --help' for usage"
     Exit 2
 }
 
@@ -50,9 +44,8 @@ Function FindJq() {
     }
     Else {
         $TempFile = [System.IO.Path]::GetTempFileName() -Replace '.tmp', '.exe'
-        DownloadFile `
-            https://github.com/jqlang/jq/releases/latest/download/jq-windows-amd64.exe `
-            $TempFile
+        Invoke-WebRequest -UseBasicParsing -OutFile $TempFile -Uri `
+            https://github.com/jqlang/jq/releases/latest/download/jq-windows-amd64.exe
         Write-Output $TempFile
     }
 }
@@ -64,7 +57,7 @@ Function FindScripts($Version) {
     $Response = Invoke-WebRequest -UseBasicParsing -Uri "$Uri"
 
     $JqBin = FindJq
-    Write-Output "$Response" | & $JqBin --raw-output "$Filter"
+    Write-Output "$Response" | & $JqBin --exit-status --raw-output "$Filter"
 }
 
 # Print log message to stdout if logging is enabled.
@@ -149,7 +142,8 @@ Function Main() {
                 $MatchFound = $True
                 Log "Installing script $Name..."
 
-                DownloadFile "$SrcPrefix/$Script.ps1" "$DestDir/$Script.ps1"
+                Invoke-WebRequest -UseBasicParsing -OutFile `
+                    "$DestDir/$Script.ps1" -Uri "$SrcPrefix/$Script.ps1"
                 Log "Installed $(& $Name --version)."
             }
         }